refactor: moved /[semester] components into /[study directory
Updated file structure of the components directory so that their path is similar to the routes in which they are used in.
