+++ app/(components)/[study]/[semester]/LectureListItem.tsx
