+++ app/(components)/[study]/[semester]/LectureListProvider.tsx
