+++ app/(components)/[study]/[semester]/LectureSearch.tsx
