+++ app/(components)/[study]/[semester]/SemesterSelection.tsx
