+++ app/(components)/[study]/[semester]/[id]/LectureDescription.tsx
