+++ app/(components)/[study]/[semester]/[id]/LectureSchedule.tsx
