+++ app/lectures/[study]/[semester]/SemesterLectures.tsx
@@ -1,6 +1,6 @@
 import useSessionData from "@/components/Auth/useSessionData"
 import { GetLecturesResponse } from "@/api/lectures/route"
-import LectureListContainer from "@/components/[semester]/LectureListContainer"
+import LectureListContainer from "@/components/[study]/[semester]/LectureListContainer"
 import { notFound, redirect } from "next/navigation"
 import { Metadata } from "next"
 import { GetStudiesResponse } from "@/api/studies/route"
